Create restart_cctv-viewer.sh
Bash script to check on total memory usage, restart cctv-viewer at a certain high-limit threshold, restart it after a certain low-limit threshold, and log it's work. It also checks to ensure cctv-viewer is running at all, launches it if not, and logs its work. It also works to make sure the log file doesn't exceed a certain size.